requirements.txt
fastapi
paddlepaddle
paddleocr
uvicorn
python-multipart
requests
numpy
streamlit
Pillow